+++ packages/eslint-plugin/src/rules/no-non-null-assertion.ts
@@ -29,33 +29,19 @@ export default createRule<[], MessageIds>({
     return {
       TSNonNullExpression(node): void {
         const suggest: TSESLint.ReportSuggestionArray<MessageIds> = [];
-        function convertTokenToOptional(
-          replacement: '?.' | '?',
-        ): TSESLint.ReportFixFunction {
-          return (fixer: TSESLint.RuleFixer): TSESLint.RuleFix | null => {
-            const operator = sourceCode.getTokenAfter(
-              node.expression,
-              isNonNullAssertionPunctuator,
-            );
-            if (operator) {
-              return fixer.replaceText(operator, replacement);
-            }
 
-            return null;
-          };
+        // it always exists in non-null assertion
+        const nonNullOperator = sourceCode.getTokenAfter(
+          node.expression,
+          isNonNullAssertionPunctuator,
+        )!;
+
+        function replaceTokenWithOptional(): TSESLint.ReportFixFunction {
+          return fixer => fixer.replaceText(nonNullOperator, '?.');
         }
-        function removeToken(): TSESLint.ReportFixFunction {
-          return (fixer: TSESLint.RuleFixer): TSESLint.RuleFix | null => {
-            const operator = sourceCode.getTokenAfter(
-              node.expression,
-              isNonNullAssertionPunctuator,
-            );
-            if (operator) {
-              return fixer.remove(operator);
-            }
 
-            return null;
-          };
+        function removeToken(): TSESLint.ReportFixFunction {
+          return fixer => fixer.remove(nonNullOperator);
         }
 
         if (
@@ -67,13 +53,21 @@ export default createRule<[], MessageIds>({
               // it is x![y]?.z
               suggest.push({
                 messageId: 'suggestOptionalChain',
-                fix: convertTokenToOptional('?.'),
+                fix: replaceTokenWithOptional(),
               });
             } else {
               // it is x!.y?.z
               suggest.push({
                 messageId: 'suggestOptionalChain',
-                fix: convertTokenToOptional('?'),
+                fix(fixer) {
+                  // x!.y?.z
+                  //   ^ punctuator
+                  const punctuator = sourceCode.getTokenAfter(nonNullOperator)!;
+                  return [
+                    fixer.remove(nonNullOperator),
+                    fixer.insertTextBefore(punctuator, '?'),
+                  ];
+                },
               });
             }
           } else {
@@ -99,7 +93,7 @@ export default createRule<[], MessageIds>({
             // it is x.y?.z!()
             suggest.push({
               messageId: 'suggestOptionalChain',
-              fix: convertTokenToOptional('?.'),
+              fix: replaceTokenWithOptional(),
             });
           } else {
             // it is x.y.z!?.()